requirements.txt
Django==4.2.0
djangorestframework==3.14.0
web3==6.0.0
python-dotenv==1.0.0
django-ratelimit==4.1.0
